Move strategy plugins to a plugins subdir (plugins/strategy).
* We're going to have savefile plugins in addition to strategy plugins
  so lets make a subdirectory for all plugins.

* Pytype doesn't seem to handle PEP-420 namespace packages yet so add an
  suds/plugins/__init__.py using the compatible pkgutil.extend_path() to
  workaround that.
