+++ suds/plugins/__init__.py
@@ -2,7 +2,7 @@
 # Copyright: (C) 2022  Toshio Kuratomi
 # License: AGPL-3.0-or-later
 # See the LICENSE file for more details
-"""Pluins implementing various strategies to progress in sudoku"""
+"""Suds plugins."""
 from pkgutil import extend_path
 
 __path__ = extend_path(__path__, __name__)
