+++ suds/plugins/meson.build
@@ -1,0 +1,9 @@
+src = ['__init__.py']
+
+python.install_sources(
+  src,
+  pure: true,
+  subdir: 'suds/plugins',
+)
+
+subdir(['strategy'])
