+++ suds/plugins/strategy/__init__.py
@@ -1,0 +1,8 @@
+# Suds -- A sudoku solver
+# Copyright: (C) 2022  Toshio Kuratomi
+# License: AGPL-3.0-or-later
+# See the LICENSE file for more details
+"""Plugins implementing various strategies to progress in sudoku"""
+from pkgutil import extend_path
+
+__path__ = extend_path(__path__, __name__)
