+++ suds/plugins/strategy/meson.build
@@ -3,5 +3,5 @@ src = ['__init__.py', 'number_in.py', 'only_in.py']
 python.install_sources(
   src,
   pure: true,
-  subdir: 'suds/strategy_plugins',
+  subdir: 'suds/plugins/strategy',
 )
